Add new generic `DetailedCell` to provide type aware sorting.
